+++ common/src/main/java/com/gregtechceu/gtceu/common/data/GTMachines.java
@@ -91,6 +91,7 @@ import static com.gregtechceu.gtceu.api.GTValues.*;
 import static com.gregtechceu.gtceu.api.pattern.Predicates.*;
 import static com.gregtechceu.gtceu.api.pattern.util.RelativeDirection.*;
 import static com.gregtechceu.gtceu.api.registry.GTRegistries.REGISTRATE;
+import static com.gregtechceu.gtceu.common.data.GCyMBlocks.*;
 import static com.gregtechceu.gtceu.common.data.GTBlocks.*;
 import static com.gregtechceu.gtceu.common.data.GTCreativeModeTabs.MACHINE;
 import static com.gregtechceu.gtceu.common.data.GTMaterials.DrillingFluid;
@@ -1397,412 +1398,6 @@ public class GTMachines {
                     .register(),
             IV, LuV) : null;
 
-
-    /////////////////////////////////////////////////
-    //********** Gregicality Multiblocks **********//
-    /////////////////////////////////////////////////
-
-    public final static MultiblockMachineDefinition LARGE_MACERATION_TOWER = REGISTRATE.multiblock("large_maceration_tower", WorkableElectricMultiblockMachine::new)
-            .langValue("Large Maceration Tower")
-            .rotationState(RotationState.NON_Y_AXIS)
-            .recipeType(GTRecipeTypes.MACERATOR_RECIPES)
-            .recipeModifier(GTRecipeModifiers.ELECTRIC_OVERCLOCK.apply(OverclockingLogic.NON_PERFECT_OVERCLOCK))
-            .appearanceBlock(CASING_SECURE_MACERATION)
-            .pattern(definition -> FactoryBlockPattern.start()
-                    .aisle("XXXXX", "XXXXX","XXXXX", "XXXXX")
-                    .aisle("XXXXX", "XGGGX","XGGGX", "X###X")
-                    .aisle("XXXXX", "XGGGX","XGGGX", "X###X")
-                    .aisle("XXXXX", "XGGGX","XGGGX", "X###X")
-                    .aisle("XXXXX", "XXXXX","XXSXX", "XXXXX")
-                    .where('S', controller(blocks(definition.get())))
-                    .where('X', blocks(CASING_SECURE_MACERATION.get()).setMinGlobalLimited(14)
-                            .or(Predicates.autoAbilities(definition.getRecipeType()))
-                            .or(Predicates.autoAbilities(true, true)))
-                    .where('G', Predicates.blocks(CRUSHING_WHEELS.get()))
-                    .where('#', Predicates.air())
-                    .build())
-            .workableCasingRenderer(GTCEu.id("block/casings/gcmb/machine_casing_secure_maceration"),
-                    GTCEu.id("block/multiblock/gcmb/large_maceration_tower"), false)
-            .compassSections(GTCompassSections.TIER[IV])
-            .compassNodeSelf()
-            .register();
-
-    public final static MultiblockMachineDefinition LARGE_ORE_WASHING_PLANT = REGISTRATE.multiblock("large_ore_wahing_plant", WorkableElectricMultiblockMachine::new)
-            .langValue("Large Ore Washing Plant")
-            .rotationState(RotationState.NON_Y_AXIS)
-            .recipeType(GTRecipeTypes.ORE_WASHER_RECIPES)
-            .recipeModifier(GTRecipeModifiers.ELECTRIC_OVERCLOCK.apply(OverclockingLogic.NON_PERFECT_OVERCLOCK))
-            .appearanceBlock(CASING_WATERTIGHT)
-            .pattern(definition -> FactoryBlockPattern.start()
-                    .aisle("XXXXX","XXXXX","XXXXX")
-                    .aisle("XXXXX","XPPPX","XWWWX")
-                    .aisle("XXXXX","XWWWX","XWWWX")
-                    .aisle("XXXXX","XWWWX","XWWWX")
-                    .aisle("XXXXX","XWWWX","XWWWX")
-                    .aisle("XXXXX","XPPPX","XWWWX")
-                    .aisle("XXXXX","XXSXX","XXXXX")
-                    .where('S', controller(blocks(definition.get())))
-                    .where('X', blocks(CASING_WATERTIGHT.get()).setMinGlobalLimited(14)
-                            .or(Predicates.autoAbilities(definition.getRecipeType()))
-                            .or(Predicates.autoAbilities(true, true)))
-                    .where('W', Predicates.blocks(Blocks.WATER))
-                    .where('P', Predicates.blocks(CASING_TITANIUM_PIPE.get()))
-                    .build())
-            .workableCasingRenderer(GTCEu.id("block/casings/gcmb/machine_casing_watertight"),
-                    GTCEu.id("block/multiblock/gcmb/large_ore_washing_plant"), false)
-            .compassSections(GTCompassSections.TIER[IV])
-            .compassNodeSelf()
-            .register();
-
-    public final static MultiblockMachineDefinition LARGE_CENTRIFUGE = REGISTRATE.multiblock("large_centrifuge", WorkableElectricMultiblockMachine::new)
-            .langValue("Large Centrifugal Unit")
-            .rotationState(RotationState.NON_Y_AXIS)
-            .recipeType(GTRecipeTypes.LARGE_CENTRIFUGE_RECIPES)
-            .recipeModifier(GTRecipeModifiers.ELECTRIC_OVERCLOCK.apply(OverclockingLogic.NON_PERFECT_OVERCLOCK))
-            .appearanceBlock(CASING_VIBRATION_SAFE)
-            .pattern(definition -> FactoryBlockPattern.start()
-                    .aisle("#XXX#","XXXXX","#XXX#")
-                    .aisle("XXXXX","X#P#X","XXXXX")
-                    .aisle("XXXXX","XP#PX","XXXXX")
-                    .aisle("XXXXX","X#P#X","XXXXX")
-                    .aisle("#XXX#","XXSXX","#XXX#")
-                    .where('S', controller(blocks(definition.get())))
-                    .where('X', blocks(CASING_VIBRATION_SAFE.get()).setMinGlobalLimited(14)
-                            .or(Predicates.autoAbilities(definition.getRecipeType()))
-                            .or(Predicates.autoAbilities(true, true)))
-                    .where('P', Predicates.blocks(CASING_STEEL_PIPE.get()))
-                    .where('#', Predicates.air())
-                    .build())
-            .workableCasingRenderer(GTCEu.id("block/casings/gcmb/machine_casing_vibration_safe"),
-                    GTCEu.id("block/multiblock/gcmb/large_centrifuge"), false)
-            .compassSections(GTCompassSections.TIER[IV])
-            .compassNodeSelf()
-            .register();
-
-    public final static MultiblockMachineDefinition LARGE_THERMAL_CENTRIFUGE = REGISTRATE.multiblock("large_thermal_centrifuge", WorkableElectricMultiblockMachine::new)
-            .langValue("Large Thermal Centrifuge")
-            .rotationState(RotationState.NON_Y_AXIS)
-            .recipeType(GTRecipeTypes.THERMAL_CENTRIFUGE_RECIPES)
-            .recipeModifier(GTRecipeModifiers.ELECTRIC_OVERCLOCK.apply(OverclockingLogic.NON_PERFECT_OVERCLOCK))
-            .appearanceBlock(CASING_VIBRATION_SAFE)
-            .pattern(definition -> FactoryBlockPattern.start()
-                    .aisle("#XXX#","XXXXX","#XXX#")
-                    .aisle("XXXXX","X#P#X","XXXXX")
-                    .aisle("XXXXX","XP#PX","XXXXX")
-                    .aisle("XXXXX","X#P#X","XXXXX")
-                    .aisle("#XXX#","XXSXX","#XXX#")
-                    .where('S', controller(blocks(definition.get())))
-                    .where('X', blocks(CASING_VIBRATION_SAFE.get()).setMinGlobalLimited(14)
-                            .or(Predicates.autoAbilities(definition.getRecipeType()))
-                            .or(Predicates.autoAbilities(true, true)))
-                    .where('P', Predicates.blocks(CASING_TITANIUM_PIPE.get()))
-                    .where('#', Predicates.air())
-                    .build())
-            .workableCasingRenderer(GTCEu.id("block/casings/gcmb/machine_casing_vibration_safe"),
-                    GTCEu.id("block/multiblock/gcmb/large_thermal_centrifuge"), false)
-            .compassSections(GTCompassSections.TIER[IV])
-            .compassNodeSelf()
-            .register();
-
-    public final static MultiblockMachineDefinition LARGE_MIXER = REGISTRATE.multiblock("large_mixer", WorkableElectricMultiblockMachine::new)
-            .langValue("Large Mixing Vessel")
-            .rotationState(RotationState.NON_Y_AXIS)
-            .recipeType(GTRecipeTypes.LARGE_MIXER_RECIPES)
-            .recipeModifier(GTRecipeModifiers.ELECTRIC_OVERCLOCK.apply(OverclockingLogic.NON_PERFECT_OVERCLOCK))
-            .appearanceBlock(CASING_VIBRATION_SAFE)
-            .pattern(definition -> FactoryBlockPattern.start()
-                    .aisle("#XXX#","#XXX#","#XXX#","#XXX#","#XXX#","##F##")
-                    .aisle("XXXXX","X#P#X","X###X","X#P#X","X###X","##F##")
-                    .aisle("XXXXX","XPPPX","X#P#X","XPPPX","X#G#X","FFGFF")
-                    .aisle("XXXXX","X#P#X","X###X","X#P#X","X###X","##F##")
-                    .aisle("#XXX#","#XSX#","#XXX#","#XXX#","#XXX#","##F##")
-                    .where('S', controller(blocks(definition.get())))
-                    .where('X', blocks(CASING_VIBRATION_SAFE.get()).setMinGlobalLimited(14)
-                            .or(autoAbilities(definition.getRecipeType()))
-                            .or(autoAbilities(true, true)))
-                    .where('F', blocks(ChemicalHelper.getBlock(TagPrefix.frameGt, GTMaterials.HastelloyX)))
-                    .where('G', blocks(CASING_STAINLESS_STEEL_GEARBOX.get()))
-                    .where('P', blocks(CASING_TITANIUM_PIPE.get()))
-                    .where('#', Predicates.air())
-                    .build())
-            .workableCasingRenderer(GTCEu.id("block/casings/gcmb/machine_casing_vibration_safe"),
-                    GTCEu.id("block/multiblock/gcmb/large_mixer"), false)
-            .compassSections(GTCompassSections.TIER[IV])
-            .compassNodeSelf()
-            .register();
-
-    public final static MultiblockMachineDefinition LARGE_ELECTROLYZER = REGISTRATE.multiblock("large_electrolyzer", WorkableElectricMultiblockMachine::new)
-            .langValue("Large Electrolysis Chamber")
-            .rotationState(RotationState.NON_Y_AXIS)
-            .recipeType(GTRecipeTypes.ELECTROLYZER_RECIPES)
-            .recipeModifier(GTRecipeModifiers.ELECTRIC_OVERCLOCK.apply(OverclockingLogic.NON_PERFECT_OVERCLOCK))
-            .appearanceBlock(CASING_NONCONDUCTING)
-            .pattern(definition -> FactoryBlockPattern.start()
-                    .aisle("XXXXX","XXXXX","XXXXX")
-                    .aisle("XXXXX","XCCCX","XCCCX")
-                    .aisle("XXXXX","XCCCX","XCCCX")
-                    .aisle("XXXXX","XXSXX","XXXXX")
-                    .where('S', controller(blocks(definition.get())))
-                    .where('X', blocks(CASING_NONCONDUCTING.get()).setMinGlobalLimited(14)
-                            .or(Predicates.autoAbilities(definition.getRecipeType()))
-                            .or(Predicates.autoAbilities(true, true)))
-                    .where('C', blocks(GTBlocks.ELECTROLYTIC_CELL.get()))
-                    .build())
-            .workableCasingRenderer(GTCEu.id("block/casings/gcmb/machine_casing_nonconducting"),
-                    GTCEu.id("block/multiblock/gcmb/large_electrolyzer"), false)
-            .compassSections(GTCompassSections.TIER[IV])
-            .compassNodeSelf()
-            .register();
-
-    public final static MultiblockMachineDefinition LARGE_PACKER = REGISTRATE.multiblock("large_packer", WorkableElectricMultiblockMachine::new)
-            .langValue("Large Packaging Machine")
-            .rotationState(RotationState.NON_Y_AXIS)
-            .recipeType(GTRecipeTypes.PACKER_RECIPES)
-            .recipeModifier(GTRecipeModifiers.ELECTRIC_OVERCLOCK.apply(OverclockingLogic.NON_PERFECT_OVERCLOCK))
-            .appearanceBlock(CASING_TUNGSTENSTEEL_ROBUST)
-            .pattern(definition -> FactoryBlockPattern.start()
-                    .aisle("XXX","XXX","XXX")
-                    .aisle("XXX","X#X","XXX")
-                    .aisle("XXX","X#X","XXX")
-                    .aisle("XXX","X#X","XXX")
-                    .aisle("XXX","X#X","XXX")
-                    .aisle("XXX","XSX","XXX")
-                    .where('S', controller(blocks(definition.get())))
-                    .where('X', blocks(CASING_TUNGSTENSTEEL_ROBUST.get()).setMinGlobalLimited(14)
-                            .or(Predicates.autoAbilities(definition.getRecipeType()))
-                            .or(Predicates.autoAbilities(true, true)))
-                    .where('#', Predicates.air())
-                    .build())
-            .workableCasingRenderer(GTCEu.id("block/casings/solid/machine_casing_robust_tungstensteel"),
-                    GTCEu.id("block/multiblock/gcmb/large_packer"), false)
-            .compassSections(GTCompassSections.TIER[HV])
-            .compassNodeSelf()
-            .register();
-
-    public final static MultiblockMachineDefinition LARGE_ASSEMBLER = REGISTRATE.multiblock("large_assembler", WorkableElectricMultiblockMachine::new)
-            .langValue("Large Assembling Factory")
-            .rotationState(RotationState.NON_Y_AXIS)
-            .recipeType(GTRecipeTypes.ASSEMBLER_RECIPES)
-            .recipeModifier(GTRecipeModifiers.ELECTRIC_OVERCLOCK.apply(OverclockingLogic.NON_PERFECT_OVERCLOCK))
-            .appearanceBlock(CASING_LARGE_SCALE_ASSEMBLING)
-            .pattern(definition -> FactoryBlockPattern.start()
-                    .aisle("XXXXXXXXX","XXXXXXXXX","XXXXXXXXX")
-                    .aisle("XXXXXXXXX","XGGGCCCCX","XXXXXXXXX")
-                    .aisle("XXXXXXXXX","XGGGXXSXX","XXXXX###X")
-                    .where('S', controller(blocks(definition.get())))
-                    .where('X', blocks(CASING_LARGE_SCALE_ASSEMBLING.get()).setMinGlobalLimited(14)
-                            .or(Predicates.autoAbilities(definition.getRecipeType()))
-                            .or(Predicates.autoAbilities(true, true)))
-                    .where('G', Predicates.blocks(CASING_TEMPERED_GLASS.get()))
-                    .where('C', Predicates.blocks(CASING_ASSEMBLY_LINE.get()))
-                    .where('#', Predicates.air())
-                    .build())
-            .workableCasingRenderer(GTCEu.id("block/casings/gcmb/machine_casing_large_scale_assembling"),
-                    GTCEu.id("block/multiblock/gcmb/large_assembler"), false)
-            .compassSections(GTCompassSections.TIER[IV])
-            .compassNodeSelf()
-            .register();
-
-    public final static MultiblockMachineDefinition LARGE_CIRCUIT_ASSEMBLER = REGISTRATE.multiblock("large_circuit_assembler", WorkableElectricMultiblockMachine::new)
-            .langValue("Large Circuit Assembling Facility")
-            .rotationState(RotationState.NON_Y_AXIS)
-            .recipeType(GTRecipeTypes.CIRCUIT_ASSEMBLER_RECIPES)
-            .recipeModifier(GTRecipeModifiers.ELECTRIC_OVERCLOCK.apply(OverclockingLogic.NON_PERFECT_OVERCLOCK))
-            .appearanceBlock(CASING_LARGE_SCALE_ASSEMBLING)
-            .pattern(definition -> FactoryBlockPattern.start()
-                    .aisle("XXXXXXXXX","XXXXXXXXX","XXXXXXXXX")
-                    .aisle("XXXXXXXXX","XGGGCCCCX","XXXXXXXXX")
-                    .aisle("XXXXXXXXX","XGGGXXSXX","XXXXX###X")
-                    .where('S', controller(blocks(definition.get())))
-                    .where('X', blocks(CASING_LARGE_SCALE_ASSEMBLING.get()).setMinGlobalLimited(14)
-                            .or(Predicates.autoAbilities(definition.getRecipeType()))
-                            .or(Predicates.autoAbilities(true, true)))
-                    .where('G', Predicates.blocks(CASING_TEMPERED_GLASS.get()))
-                    .where('C', Predicates.blocks(CASING_ASSEMBLY_LINE.get()))
-                    .where('#', Predicates.air())
-                    .build())
-            .workableCasingRenderer(GTCEu.id("block/casings/gcmb/machine_casing_large_scale_assembling"),
-                    GTCEu.id("block/multiblock/gcmb/large_circuit_assembler"), false)
-            .compassSections(GTCompassSections.TIER[IV])
-            .compassNodeSelf()
-            .register();
-
-    public final static MultiblockMachineDefinition LARGE_ARC_SMELTER = REGISTRATE.multiblock("large_arc_smelter", WorkableElectricMultiblockMachine::new)
-            .langValue("Large Arc Smelter")
-            .rotationState(RotationState.NON_Y_AXIS)
-            .recipeType(GTRecipeTypes.ARC_FURNACE_RECIPES)
-            .recipeModifier(GTRecipeModifiers.ELECTRIC_OVERCLOCK.apply(OverclockingLogic.NON_PERFECT_OVERCLOCK))
-            .appearanceBlock(CASING_HIGH_TEMPERATURE_SMELTING)
-            .pattern(definition -> FactoryBlockPattern.start()
-                    .aisle("#XXX#","#XXX#","#XXX#","#XXX#")
-                    .aisle("XXXXX","XC#CX","XC#CX","XXXXX")
-                    .aisle("XXXXX","X###X","X###X","XXXXX")
-                    .aisle("XXXXX","X#C#X","X#C#X","XXXXX")
-                    .aisle("#XXX#","#XSX#","#XXX#","#XXX#")
-                    .where('S', controller(blocks(definition.get())))
-                    .where('X', blocks(CASING_HIGH_TEMPERATURE_SMELTING.get()).setMinGlobalLimited(14)
-                            .or(Predicates.autoAbilities(definition.getRecipeType()))
-                            .or(Predicates.autoAbilities(true, true)))
-                    .where('C', Predicates.blocks(GTBlocks.MOLYBDENUM_DISILICIDE_COIL_BLOCK.get()))
-                    .where('#', Predicates.air())
-                    .build())
-            .workableCasingRenderer(GTCEu.id("block/casings/gcmb/machine_casing_high_temperature_smelting"),
-                    GTCEu.id("block/multiblock/gcmb/large_arc_smelter"), false)
-            .compassSections(GTCompassSections.TIER[IV])
-            .compassNodeSelf()
-            .register();
-
-    public final static MultiblockMachineDefinition LARGE_ENGRAVING_LASER = REGISTRATE.multiblock("large_engraving_laser", WorkableElectricMultiblockMachine::new)
-            .langValue("Large Engraving Laser")
-            .rotationState(RotationState.NON_Y_AXIS)
-            .recipeType(GTRecipeTypes.LARGE_ENGRAVER_RECIPES)
-            .recipeModifier(GTRecipeModifiers.ELECTRIC_OVERCLOCK.apply(OverclockingLogic.NON_PERFECT_OVERCLOCK))
-            .appearanceBlock(CASING_LASER_SAFE_ENGRAVING)
-            .pattern(definition -> FactoryBlockPattern.start()
-                    .aisle("XXXXX","XXGXX","XXGXX","XXXXX")
-                    .aisle("XXXXX","X###X","X###X","XKKKX")
-                    .aisle("XXXXX","G###G","G#C#G","XKXKX")
-                    .aisle("XXXXX","X###X","X###X","XKKKX")
-                    .aisle("XXSXX","XXGXX","XXGXX","XXXXX")
-                    .where('S', controller(blocks(definition.get())))
-                    .where('C', blocks(CASING_TUNGSTENSTEEL_PIPE.get()))
-                    .where('X', blocks(CASING_LASER_SAFE_ENGRAVING.get()).setMinGlobalLimited(14)
-                            .or(Predicates.autoAbilities(definition.getRecipeType()))
-                            .or(Predicates.autoAbilities(true, true)))
-                    .where('G', blocks(CASING_TEMPERED_GLASS.get()))
-                    .where('K', blocks(CASING_GRATE.get()))
-                    .where('#', Predicates.air())
-                    .build())
-            .workableCasingRenderer(GTCEu.id("block/casings/gcmb/machine_casing_laser_safe_engraving"),
-                    GTCEu.id("block/multiblock/gcmb/large_engraving_laser"), false)
-            .compassSections(GTCompassSections.TIER[IV])
-            .compassNodeSelf()
-            .register();
-
-    public final static MultiblockMachineDefinition LARGE_SIFTING_FUNNEL = REGISTRATE.multiblock("large_sifting_funnel", WorkableElectricMultiblockMachine::new)
-            .langValue("Large Sifting Funnel")
-            .rotationState(RotationState.NON_Y_AXIS)
-            .recipeType(GTRecipeTypes.SIFTER_RECIPES)
-            .recipeModifier(GTRecipeModifiers.ELECTRIC_OVERCLOCK.apply(OverclockingLogic.NON_PERFECT_OVERCLOCK))
-            .appearanceBlock(CASING_VIBRATION_SAFE)
-            .pattern(definition -> FactoryBlockPattern.start()
-                    .aisle("#X#X#","#X#X#","#XXX#","XXXXX","#XXX#")
-                    .aisle("XXXXX","X#X#X","XKKKX","XKKKX","X###X")
-                    .aisle("#XXX#","#X#X#","XKKKX","XKKKX","X###X")
-                    .aisle("XXXXX","X#X#X","XKKKX","XKKKX","X###X")
-                    .aisle("#X#X#","#X#X#","#XSX#","XXXXX","#XXX#")
-                    .where('S', controller(blocks(definition.get())))
-                    .where('X', blocks(CASING_VIBRATION_SAFE.get()).setMinGlobalLimited(14)
-                            .or(Predicates.autoAbilities(definition.getRecipeType()))
-                            .or(Predicates.autoAbilities(true, true)))
-                    .where('K', blocks(CASING_GRATE.get()))
-                    .where('#', Predicates.air())
-                    .build())
-            .workableCasingRenderer(GTCEu.id("block/casings/gcmb/machine_casing_vibration_safe"),
-                    GTCEu.id("block/multiblock/gcmb/large_sifting_funnel"), false)
-            .compassSections(GTCompassSections.TIER[IV])
-            .compassNodeSelf()
-            .register();
-
-    public final static MultiblockMachineDefinition BLAST_ALLOY_SMELTER = REGISTRATE.multiblock("blast_alloy_smelter", WorkableElectricMultiblockMachine::new)
-            .langValue("Blast Alloy Smelter")
-            .rotationState(RotationState.NON_Y_AXIS)
-            .recipeType(GTRecipeTypes.BLAST_ALLOY_RECIPES)
-            .recipeModifier(GTRecipeModifiers::multiSmelterOverclock)
-            .appearanceBlock(CASING_HIGH_TEMPERATURE_SMELTING)
-            .pattern(definition -> FactoryBlockPattern.start()
-                    .aisle("#XXX#", "#CCC#", "#GGG#", "#CCC#", "#XXX#")
-                    .aisle("XXXXX", "C###C", "G###G", "C###C", "XXXXX")
-                    .aisle("XXXXX", "C###C", "G###G", "C###C", "XXMXX")
-                    .aisle("XXXXX", "C###C", "G###G", "C###C", "XXXXX")
-                    .aisle("#XSX#", "#CCC#", "#GGG#", "#CCC#", "#XXX#")
-                    .where('S', controller(blocks(definition.get())))
-                    .where('X', blocks(CASING_HIGH_TEMPERATURE_SMELTING.get()).setMinGlobalLimited(14)
-                            .or(autoAbilities(definition.getRecipeType()))
-                            .or(autoAbilities(true, true)))
-                    .where('C', heatingCoils())
-                    .where('M', abilities(PartAbility.MUFFLER))
-                    .where('G', blocks(HEAT_VENT.get()))
-                    .where('#', air())
-                    .build())
-            .workableCasingRenderer(GTCEu.id("block/casings/gcmb/machine_casing_high_temperature_smelting"),
-                    GTCEu.id("block/multiblock/gcmb/blast_alloy_smelter"), false)
-            .compassSections(GTCompassSections.TIER[IV])
-            .compassNodeSelf()
-            .register();
-
-    public final static MultiblockMachineDefinition MEGA_BLAST_FURNACE = REGISTRATE.multiblock("mega_blast_furnace", WorkableElectricMultiblockMachine::new)
-            .langValue("Rotary Hearth Furnace")
-            .rotationState(RotationState.NON_Y_AXIS)
-            .recipeType(GTRecipeTypes.BLAST_RECIPES)
-            .recipeModifier(GTRecipeModifiers::multiSmelterOverclock)
-            .appearanceBlock(CASING_HIGH_TEMPERATURE_SMELTING)
-            .pattern(definition -> FactoryBlockPattern.start()
-                    .aisle("##XXXXXXXXX##", "##XXXXXXXXX##", "#############", "#############", "#############", "#############", "#############", "#############", "#############", "#############", "#############", "#############", "#############", "#############", "#############", "#############", "#############")
-                    .aisle("#XXXXXXXXXXX#", "#XXXXXXXXXXX#", "###F#####F###", "###F#####F###", "###FFFFFFF###", "#############", "#############", "#############", "#############", "#############", "####FFFFF####", "#############", "#############", "#############", "#############", "#############", "#############")
-                    .aisle("XXXXXXXXXXXXX", "XXXXVVVVVXXXX", "##F#######F##", "##F#######F##", "##FFFHHHFFF##", "##F#######F##", "##F#######F##", "##F#######F##", "##F#######F##", "##F#######F##", "##FFFHHHFFF##", "#############", "#############", "#############", "#############", "#############", "###TTTTTTT###")
-                    .aisle("XXXXXXXXXXXXX", "XXXXXXXXXXXXX", "#F####P####F#", "#F####P####F#", "#FFHHHPHHHFF#", "######P######", "######P######", "######P######", "######P######", "######P######", "##FHHHPHHHF##", "######P######", "######P######", "######P######", "######P######", "######P######", "##TTTTPTTTT##")
-                    .aisle("XXXXXXXXXXXXX", "XXVXXXXXXXVXX", "####BBPBB####", "####TITIT####", "#FFHHHHHHHFF#", "####BITIB####", "####CCCCC####", "####CCCCC####", "####CCCCC####", "####BITIB####", "#FFHHHHHHHFF#", "####BITIB####", "####CCCCC####", "####CCCCC####", "####CCCCC####", "####BITIB####", "##TTTTPTTTT##")
-                    .aisle("XXXXXXXXXXXXX", "XXVXXXXXXXVXX", "####B###B####", "####I###I####", "#FHHH###HHHF#", "####I###I####", "####C###C####", "####C###C####", "####C###C####", "####I###I####", "#FHHH###HHHF#", "####I###I####", "####C###C####", "####C###C####", "####C###C####", "####I###I####", "##TTTTPTTTT##")
-                    .aisle("XXXXXXXXXXXXX", "XXVXXXXXXXVXX", "###PP###PP###", "###PT###TP###", "#FHPH###HPHF#", "###PT###TP###", "###PC###CP###", "###PC###CP###", "###PC###CP###", "###PT###TP###", "#FHPH###HPHF#", "###PT###TP###", "###PC###CP###", "###PC###CP###", "###PC###CP###", "###PT###TP###", "##TPPPMPPPT##")
-                    .aisle("XXXXXXXXXXXXX", "XXVXXXXXXXVXX", "####B###B####", "####I###I####", "#FHHH###HHHF#", "####I###I####", "####C###C####", "####C###C####", "####C###C####", "####I###I####", "#FHHH###HHHF#", "####I###I####", "####C###C####", "####C###C####", "####C###C####", "####I###I####", "##TTTTPTTTT##")
-                    .aisle("XXXXXXXXXXXXX", "XXVXXXXXXXVXX", "####BBPBB####", "####TITIT####", "#FFHHHHHHHFF#", "####BITIB####", "####CCCCC####", "####CCCCC####", "####CCCCC####", "####BITIB####", "#FFHHHHHHHFF#", "####BITIB####", "####CCCCC####", "####CCCCC####", "####CCCCC####", "####BITIB####", "##TTTTPTTTT##")
-                    .aisle("XXXXXXXXXXXXX", "XXXXXXXXXXXXX", "#F####P####F#", "#F####P####F#", "#FFHHHPHHHFF#", "######P######", "######P######", "######P######", "######P######", "######P######", "##FHHHPHHHF##", "######P######", "######P######", "######P######", "######P######", "######P######", "##TTTTPTTTT##")
-                    .aisle("XXXXXXXXXXXXX", "XXXXVVVVVXXXX", "##F#######F##", "##F#######F##", "##FFFHHHFFF##", "##F#######F##", "##F#######F##", "##F#######F##", "##F#######F##", "##F#######F##", "##FFFHHHFFF##", "#############", "#############", "#############", "#############", "#############", "###TTTTTTT###")
-                    .aisle("#XXXXXXXXXXX#", "#XXXXXXXXXXX#", "###F#####F###", "###F#####F###", "###FFFFFFF###", "#############", "#############", "#############", "#############", "#############", "####FFFFF####", "#############", "#############", "#############", "#############", "#############", "#############")
-                    .aisle("##XXXXXXXXX##", "##XXXXSXXXX##", "#############", "#############", "#############", "#############", "#############", "#############", "#############", "#############", "#############", "#############", "#############", "#############", "#############", "#############", "#############")
-                    .where('S', controller(blocks(definition.get())))
-                    .where('X', blocks(CASING_HIGH_TEMPERATURE_SMELTING.get()).setMinGlobalLimited(14)
-                            .or(autoAbilities(definition.getRecipeType()))
-                            .or(autoAbilities(true, true)))
-                    .where('C', heatingCoils())
-                    .where('M', abilities(PartAbility.MUFFLER))
-                    .where('F', blocks(ChemicalHelper.getBlock(TagPrefix.frameGt,NaquadahAlloy)))
-                    .where('H', blocks(CASING_HIGH_TEMPERATURE_SMELTING.get()))
-                    .where('T', blocks(CASING_TUNGSTENSTEEL_ROBUST.get()))
-                    .where('B', blocks(FIREBOX_TUNGSTENSTEEL.get()))
-                    .where('P', blocks(CASING_TUNGSTENSTEEL_PIPE.get()))
-                    .where('I', blocks(CASING_EXTREME_ENGINE_INTAKE.get()))
-                    .where('V', blocks(HEAT_VENT.get()))
-                    .where('M', abilities(PartAbility.MUFFLER))
-                    .where('#', air())
-                    .build())
-            .workableCasingRenderer(GTCEu.id("block/casings/gcmb/machine_casing_high_temperature_smelting"),
-                    GTCEu.id("block/multiblock/gcmb/mega_blast_furnace"), false)
-            .compassSections(GTCompassSections.TIER[LuV])
-            .compassNodeSelf()
-            .register();
-
-    public final static MultiblockMachineDefinition MEGA_VACUUM_FREEZER = REGISTRATE.multiblock("mega_vacuum_freezer", WorkableElectricMultiblockMachine::new)
-            .langValue("Bulk Blast Chiller")
-            .rotationState(RotationState.NON_Y_AXIS)
-            .recipeType(GTRecipeTypes.VACUUM_RECIPES)
-            .recipeModifier(GTRecipeModifiers.ELECTRIC_OVERCLOCK.apply(OverclockingLogic.PERFECT_OVERCLOCK))
-            .appearanceBlock(CASING_ALUMINIUM_FROSTPROOF)
-            .pattern(definition -> FactoryBlockPattern.start()
-                    .aisle("XXXXXXX#KKK", "XXXXXXX#KVK", "XXXXXXX#KVK", "XXXXXXX#KVK", "XXXXXXX#KKK", "XXXXXXX####", "XXXXXXX####")
-                    .aisle("XXXXXXX#KVK", "XPPPPPPPPPV", "XP#P#PX#VPV", "XPPPPPPPPPV", "XP#P#PX#KVK", "XPPPPPX####", "XXXXXXX####")
-                    .aisle("XXXXXXX#KVK", "XP#P#PX#VPV", "X#####X#VPV", "XP###PX#VPV", "X#####X#KVK", "XP#P#PX####", "XXXXXXX####")
-                    .aisle("XXXXXXX#KVK", "XP#P#PPPPPV", "X#####X#VPV", "XP###PPPPPV", "X#####X#KVK", "XP#P#PX####", "XXXXXXX####")
-                    .aisle("XXXXXXX#KKK", "XPPPPPX#KVK", "XP###PX#KVK", "XP###PX#KVK", "XP###PX#KKK", "XPPPPPX####", "XXXXXXX####")
-                    .aisle("#XXXXX#####", "#XXSXX#####", "#XGGGX#####", "#XGGGX#####", "#XGGGX#####", "#XXXXX#####", "###########")
-                    .where('S', controller(blocks(definition.get())))
-                    .where('X', blocks(CASING_ALUMINIUM_FROSTPROOF.get()).setMinGlobalLimited(140)
-                            .or(autoAbilities(definition.getRecipeType()))
-                            .or(autoAbilities(true, false)))
-                    .where('G', blocks(CASING_TEMPERED_GLASS.get()))
-                    .where('K', blocks(CASING_STAINLESS_CLEAN.get()))
-                    .where('P', blocks(CASING_TUNGSTENSTEEL_PIPE.get()))
-                    .where('V', blocks(HEAT_VENT.get()))
-                    .where('#', air())
-                    .build())
-            .workableCasingRenderer(GTCEu.id("block/casings/solid/machine_casing_frost_proof"),
-                    GTCEu.id("block/multiblock/gcmb/mega_vacuum_freezer"), false)
-            .compassSections(GTCompassSections.TIER[LuV])
-            .compassNodeSelf()
-            .register();
-
-
     //////////////////////////////////////
     //**********     Misc     **********//
     //////////////////////////////////////
@@ -2114,6 +1709,7 @@ public class GTMachines {
     }
 
     public static void init() {
+        GCyMMachines.init();
         if (GTCEu.isCreateLoaded()) {
             new GTCreateMachines();
         }
