+++ common/src/main/java/com/gregtechceu/gtceu/data/recipe/misc/MetaTileEntityLoader.java
@@ -212,21 +212,6 @@ public class MetaTileEntityLoader {
 
         VanillaRecipeHelper.addShapedRecipe(provider, true, "large_chemical_reactor", GTMachines.LARGE_CHEMICAL_REACTOR.asStack(), "CRC", "PMP", "CHC", 'C', CustomTags.HV_CIRCUITS, 'R', ChemicalHelper.get(TagPrefix.rotor, GTMaterials.StainlessSteel), 'P', ChemicalHelper.get(TagPrefix.pipeLargeFluid, GTMaterials.Polytetrafluoroethylene), 'M', GTItems.ELECTRIC_MOTOR_HV.asStack(), 'H', GTMachines.HULL[HV].asStack());
 
-
-        //Gregicality Multiblock Controllers
-        VanillaRecipeHelper.addShapedRecipe(provider, true, "large_macerator", GTMachines.LARGE_MACERATION_TOWER.asStack(), "PCP", "BXB", "MKM", 'C', CustomTags.IV_CIRCUITS, 'P', ChemicalHelper.get(TagPrefix.plate, GTMaterials.TungstenCarbide), 'B', GTItems.ELECTRIC_PISTON_IV.asStack(), 'M', GTItems.ELECTRIC_MOTOR_IV.asStack(), 'X', GTMachines.MACERATOR[GTValues.IV - 1].asStack(), 'K', new UnificationEntry(TagPrefix.cableGtSingle, GTMaterials.Platinum));
-        VanillaRecipeHelper.addShapedRecipe(provider, true, "large_arc_smelter", GTMachines.LARGE_ARC_SMELTER.asStack(), "KDK", "CXC", "PPP", 'C', CustomTags.IV_CIRCUITS, 'P', ChemicalHelper.get(TagPrefix.plate, GTMaterials.TantalumCarbide), 'X', GTMachines.ARC_FURNACE[GTValues.IV - 1].asStack(), 'D',ChemicalHelper.get(TagPrefix.dust, GTMaterials.Graphite) ,'K', new UnificationEntry(TagPrefix.cableGtSingle, GTMaterials.Platinum));
-        VanillaRecipeHelper.addShapedRecipe(provider, true, "large_ore_washer", GTMachines.LARGE_ORE_WASHING_PLANT.asStack(), "PGP", "CXC", "MKM", 'C', CustomTags.IV_CIRCUITS, 'G',GTBlocks.CASING_TEMPERED_GLASS.asStack() ,'P', GTItems.ELECTRIC_PUMP_IV.asStack(), 'M', GTItems.CONVEYOR_MODULE_IV.asStack(), 'X', GTMachines.ORE_WASHER[GTValues.IV - 1].asStack(), 'K', new UnificationEntry(TagPrefix.cableGtSingle, GTMaterials.Platinum));
-        VanillaRecipeHelper.addShapedRecipe(provider, true, "large_sifter", GTMachines.LARGE_SIFTING_FUNNEL.asStack(), "PCP", "EXE", "PKP", 'C', CustomTags.IV_CIRCUITS, 'P', ChemicalHelper.get(TagPrefix.plate, GTMaterials.HSLASteel), 'E', GTItems.ELECTRIC_PISTON_IV.asStack(), 'X', GTMachines.SIFTER[GTValues.IV - 1].asStack(), 'K', new UnificationEntry(TagPrefix.cableGtSingle, GTMaterials.Platinum));
-        VanillaRecipeHelper.addShapedRecipe(provider, true, "large_engraver", GTMachines.LARGE_ENGRAVING_LASER.asStack(), "ICI", "EXE", "PKP", 'C', CustomTags.IV_CIRCUITS, 'P', ChemicalHelper.get(TagPrefix.plateDouble, GTMaterials.TantalumCarbide),'I',GTItems.EMITTER_IV.asStack() ,'E', GTItems.ELECTRIC_PISTON_IV.asStack(), 'X', GTMachines.LASER_ENGRAVER[GTValues.IV - 1].asStack(), 'K', new UnificationEntry(TagPrefix.cableGtSingle, GTMaterials.Platinum));
-        VanillaRecipeHelper.addShapedRecipe(provider, true, "large_packer", GTMachines.LARGE_PACKER.asStack(), "RCR", "PXP", "KPK", 'C', CustomTags.EV_CIRCUITS, 'P', ChemicalHelper.get(TagPrefix.plate, GTMaterials.HSLASteel),'R',GTItems.ROBOT_ARM_HV.asStack() ,'K', GTItems.CONVEYOR_MODULE_HV.asStack(), 'X', GTMachines.PACKER[GTValues.HV - 1].asStack());
-        VanillaRecipeHelper.addShapedRecipe(provider, true, "large_mixer", GTMachines.LARGE_MIXER.asStack(), "FCF", "RXR", "MKM", 'C', CustomTags.IV_CIRCUITS, 'F', ChemicalHelper.get(TagPrefix.pipeNormalFluid, GTMaterials.Polybenzimidazole),'R',ChemicalHelper.get(TagPrefix.rotor, GTMaterials.Osmiridium) ,'M', GTItems.ELECTRIC_MOTOR_IV.asStack(), 'X', GTMachines.MIXER[GTValues.IV - 1].asStack(), 'K', new UnificationEntry(TagPrefix.cableGtSingle, GTMaterials.Platinum));
-        VanillaRecipeHelper.addShapedRecipe(provider, true, "large_centrifuge", GTMachines.LARGE_CENTRIFUGE.asStack(), "SFS", "CXC", "MKM", 'C', CustomTags.IV_CIRCUITS, 'F', ChemicalHelper.get(TagPrefix.pipeHugeFluid, GTMaterials.StainlessSteel),'S',ChemicalHelper.get(TagPrefix.spring, GTMaterials.MolybdenumDisilicide) ,'M', GTItems.ELECTRIC_MOTOR_IV.asStack(), 'X', GTMachines.CENTRIFUGE[GTValues.IV - 1].asStack(), 'K', new UnificationEntry(TagPrefix.cableGtSingle, GTMaterials.Platinum));
-        VanillaRecipeHelper.addShapedRecipe(provider, true, "large_thermal_centrifuge", GTMachines.LARGE_THERMAL_CENTRIFUGE.asStack(), "SFS", "CXC", "MKM", 'C', CustomTags.IV_CIRCUITS, 'F', ChemicalHelper.get(TagPrefix.pipeHugeFluid, GTMaterials.StainlessSteel),'S',ChemicalHelper.get(TagPrefix.spring, GTMaterials.MolybdenumDisilicide) ,'M', GTItems.ELECTRIC_MOTOR_IV.asStack(), 'X', GTMachines.THERMAL_CENTRIFUGE[GTValues.IV - 1].asStack(), 'K', new UnificationEntry(TagPrefix.cableGtSingle, GTMaterials.Platinum));
-        VanillaRecipeHelper.addShapedRecipe(provider, true, "large_assembler", GTMachines.LARGE_ASSEMBLER.asStack(), "RKR", "CXC", "MKM", 'C', CustomTags.IV_CIRCUITS, 'R', GTItems.ROBOT_ARM_IV.asStack() ,'M', GTItems.CONVEYOR_MODULE_IV.asStack(), 'X', GTMachines.ASSEMBLER[GTValues.IV - 1].asStack(), 'K', new UnificationEntry(TagPrefix.cableGtSingle, GTMaterials.Platinum));
-        VanillaRecipeHelper.addShapedRecipe(provider, true, "large_circuit_assembler", GTMachines.LARGE_CIRCUIT_ASSEMBLER.asStack(), "RKR", "CXC", "MKM", 'C', CustomTags.IV_CIRCUITS, 'R', GTItems.ROBOT_ARM_IV.asStack() ,'M', GTItems.CONVEYOR_MODULE_IV.asStack(), 'X', GTMachines.CIRCUIT_ASSEMBLER[GTValues.IV - 1].asStack(), 'K', new UnificationEntry(TagPrefix.cableGtSingle, GTMaterials.Platinum));
-        VanillaRecipeHelper.addShapedRecipe(provider, true, "large_electrolyzer", GTMachines.LARGE_ELECTROLYZER.asStack(), "PCP", "WXW", "PKP", 'C', CustomTags.IV_CIRCUITS, 'P', ChemicalHelper.get(TagPrefix.plate, GTMaterials.BlackSteel) ,'W', ChemicalHelper.get(TagPrefix.wireGtQuadruple, GTMaterials.Osmium), 'X', GTMachines.ELECTROLYZER[GTValues.IV - 1].asStack(), 'K', new UnificationEntry(TagPrefix.cableGtSingle, GTMaterials.Platinum));
-
         if (ConfigHolder.INSTANCE.machines.doProcessingArray) {
             VanillaRecipeHelper.addShapedRecipe(provider, true, "processing_array", GTMachines.PROCESSING_ARRAY[GTValues.IV].asStack(), "COC", "RHR", "CPC", 'C', CustomTags.IV_CIRCUITS, 'O', GTItems.TOOL_DATA_ORB.asStack(), 'R', GTItems.ROBOT_ARM_EV.asStack(), 'P', GTBlocks.FLUID_PIPE_BLOCKS.get(TagPrefix.pipeLargeFluid, GTMaterials.StainlessSteel), 'H', GTMachines.HULL[GTValues.EV].asStack());
             VanillaRecipeHelper.addShapedRecipe(provider, true, "advanced_processing_array", GTMachines.PROCESSING_ARRAY[GTValues.LuV].asStack(), "RCR", "SPE", "HNH", 'R', GTItems.ROBOT_ARM_LuV.asStack(), 'C', CustomTags.ZPM_CIRCUITS, 'S', GTItems.SENSOR_LuV, 'P', GTMachines.PROCESSING_ARRAY[IV].asStack(), 'E', GTItems.EMITTER_LuV.asStack(), 'H', new UnificationEntry(TagPrefix.plate, GTMaterials.HSSE), 'N', new UnificationEntry(TagPrefix.pipeLargeFluid, GTMaterials.Naquadah));
